requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch==2.4.1
torchvision==0.19.1
numpy==1.26.4
pillow==10.4.0
python-multipart==0.0.9